refactor: move `constants.py` to `utils`
